merge

+++ web/src/main/java/com/sacc/ComprehensiveSystem/admin/sys/entity/auto/_SysUserRole.java
@@ -20,30 +20,30 @@ public class _SysUserRole extends BasicEntity {
     // ---------------------------------
     // user_id的属性
     // ---------------------------------
-    private Integer userId;
+    private Long userId;
 
     // ---------------------------------
     // role_id的属性
     // ---------------------------------
-    private Integer roleId;
+    private Long roleId;
 
 
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return this.userId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return this.roleId;
     }
 
 
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId =  userId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId =  roleId;
     }
